pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.10",
    "google-adk>=1.5.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "PyMySQL>=1.1.0",
    "aiomysql>=0.2.0",
    "uvicorn>=0.35.0",
    "aiohttp>=3.9.0",
]
